Initial commit
I think this can be done in Rust. The library seems to support all that
I need it to and with the Rust compiler being installed on my home
directory on the grading server, it looks like this could work. The
library is a little thin on documentation, but it works for now. Could
use some more examples. Maybe I'll contribute back to the library after
this project with some lessons learned. I'm excited to be able to use
Rust for a school project now!